Completely ignore 'Ignored' columns
We marked some columns as ignored but we still trimmed,
column mapped and applied converters to them which could
trigger warnings on a column that should be completely
ignored. We've ordered the import processing steps so that
the ignored mapper occurs sooner (before column name mapping,
and trimming). So it should be both more correct and faster